重构 k8s 后台部分代码
